% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToDiscreteTime.R
\name{mapToDiscreteTime}
\alias{mapToDiscreteTime}
\title{Map Irregular Data to a Discrete Time Grid via Interpolation}
\usage{
mapToDiscreteTime(data, xvar = "value", timevar = "t", start, end, dt,
  method = "constant", rule = 2, ...)
}
\arguments{
\item{data}{A data frame}

\item{xvar}{Character string indicating the column name in \code{data}
that contains the measurements}

\item{timevar}{Character string indicating the column name in \code{data}
that contains the (numeric) time values associated with the measurements
in \code{xvar}.}

\item{start}{Numeric value of the starting time over which the discrete time data
will be generated. Defaults to \code{min(data$xvar)}.}

\item{end}{Numeric value of the ending time over which the discrete time data
will be generated. Defaults to \code{max(data$xvar)}.}

\item{dt}{Numeric value representing the time increment of measurements over the
time interval given by \code{start} and \code{end}. Must satisfy
\code{dt <= (end - start)}.}

\item{method}{Interpolation method passed to \code{approx()}. Defaults to
\code{"constant"}. See \code{help(approx)} for details.}

\item{rule}{Rule by which observations outside the interval
\code{[min(data$xvar), max(data$xvar)]} are handled. Defaults to \code{rule=2};
see \code{help(approx)} for details.}

\item{...}{Other parameters passed to \code{approx()}.}
}
\value{
A data frame containing equally-spaced, interpolated values of \code{xvar}
  over the specified time grid. Other variables in \code{data} are returned (assuming)
}
\description{
This function takes irregularly-spaced observations in time and,
  through a call to \code{approx}, interpolates the data at a user-specified,
  fixed grid of equally-spaced time points. The function is essentially a
  wrapper to \code{approx} that allows data frames as input (as opposed to
  individual vectors for the x variable and time variable). As such, it
  provides a useful framework for grouped data operations such as
  \code{plyr::ddply}.
}
\details{
The function ignores columns in \code{data} other than \code{xvar} and
  \code{timevar}.
}
\examples{
irregularData <- tibble(value=1:10, t=c(1,4,5,7,8,10,11,13,17,21))
mapToDiscreteTime(irregularData, end=24, dt=2)
mapToDiscreteTime(irregularData, end=24, dt=0.5)
}
